backend/requirements.txt
fastapi
uvicorn
openai
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
